backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.9.2
pdfplumber==0.11.4
PyPDF2==3.0.1
pillow==11.0.0
python-multipart==0.0.12
scikit-learn==1.5.2
pandas==2.2.3
openai==1.51.2
python-dotenv==1.0.1


